Extract out more general parts, for future use by other plugins, and for a central place for adding conventions.
