+++ src/main/groovy/com/trigonic/gradle/plugins/packaging/SystemPackagingExtension.groovy
@@ -1,0 +1,4 @@
+package com.trigonic.gradle.plugins.packaging
+
+class SystemPackagingExtension {
+}
