+++ src/main/resources/META-INF/gradle-plugins/os-package.properties
@@ -1,0 +1,1 @@
+implementation-class=com.trigonic.gradle.plugins.rpm.RpmPlugin
